fix # 注册（利用自己编写数据库插入方法）

+++ src/main/java/cn/buu/smart_cube/common/service/CommonService.java
@@ -3,6 +3,6 @@ package cn.buu.smart_cube.common.service;
 
 public interface CommonService {
 
-	int getOnlyKey();
+	long getOnlyKey();
 
 }
